argo_poll: Remove read_lock
read_lock is a problem because __pollwait needs to allocate memory and
may sleep.  It shouldn't be necessary since we aren't manipulating the
ring_list.  Remove ring_list and replace with locking on the
prending_recv_list/lock and ring->lock when checking the respective
protected variables.

Therre could be something subtle here where read_lock prevents a race
with write_lock.  However, poll is holding an open file descriptor, so
the backing datastructures should not go away.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>